Resolve circular import
